+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Stuck last replication error, [PR-393](https://github.com/reductstore/reductstore/pull/393)
 
+### Changed
+
+- reduct-rs: Move `reduct-rs` to separated repository, [PR-395](https://github.com/reductstore/reductstore/pull/395)
+
 ## [1.8.0] - 2024-01-24
 
 ### Added